chore(install):fixes native modules rebuild, split install and rebuild into separate files
